feat: remove 'monthly' button
this commit addresses #11 by removing the 'Monthly' button.
an unforseen consequence of this is that by selecting a year,
then clicking the 'Cumulative' button, one can view the
cumulative reporting up to that year, so clicking on the link
for 2016, then on the 'Cumulative' button will show the
cumulative reporting up to and including 2016.

see: #11